+++ lib/spack/spack/package.py
@@ -49,7 +49,7 @@ from llnl.util.filesystem import (
 from llnl.util.symlink import symlink
 
 from spack.build_environment import MakeExecutable
-from spack.builder import BaseBuilder, Builder, register_builder
+from spack.builder import BaseBuilder, Builder, BuilderWithDefaults, register_builder
 from spack.config import determine_number_of_jobs
 from spack.deptypes import ALL_TYPES as all_deptypes
 from spack.directives import (
@@ -226,6 +226,7 @@ __all__ = [
     "pwd",
     "tty",
     "Builder",
+    "BuilderWithDefaults",
     "PackageBase",
     "register_builder",
 ]